arm64: tegra186: nvs: Simplify nvs sensor definitions
Sensor Kconfig definitions are simple and straightforward, but multiple
entires in defconfig are required to run a single sensor.  This change
is intended to simplify that process such that a single defconfig entry
will automatically enable the prerequisite Kconfig entries necessary to
build the sensor drivers.

Bug 2115403

Change-Id: I1375f96e19ecfe18f48db1efb5231e65f7e08b7a
Signed-off-by: Robert Collins <rcollins@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1710797
Reviewed-by: Erik Lilliebjerg <elilliebjerg@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>