cpufreq: create a single policy
create a single policy for all 8 CPUs, so that the frequency
change is in lock-step for all CPUs.

This is controlled by property for device tree node "cpufreq":
cpufreq_single_policy. When this property is present, a single
policy is created. If not, one policy is created for each cluster.

Bug 2106745

Change-Id: Idfba13acfdb6c73b368e463d731f008ffc7abd49
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1710748
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>