thermal: pwn-fan: add pwm polarity support.
When we calculate the duty cycle with the given
pwm target, we consider the pwm polarity is inversed.
However, the polarity of pwm fan is different on
different board.

Add a property in the pwm fan driver and dts to show
the polarity of the related pwm fan since not every
pwm chip implements the polarity related APIs in kernel
pwm frameworks.

Bug 200412263

Change-Id: Ic769f5bdd9030bc41271b12f85c6423700c40c23
Signed-off-by: Darren Sun <darrens@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1711965
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Cyril Raju <craju@nvidia.com>
Reviewed-by: Daniel Fu <danifu@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>