dts: t19x: common: fix sensor readout_orientation
The readout_orientation field in sensor DTs files determines the
ANDROID_SENSOR_ORIENTATION attribute, which is used by the Android
Camera2 and SHIELD camera apps to determine the rotation/orientation
of the preview and captured image.

Historically, this was set to 90 degrees CW on ov5693 sensors for the
Tegra SHIELD tablets and accompanying development boards, the value
has erroneously propagated to imx274 configs. On systems w/o a
built-in display (default Portrait orientation), the sensor image is
improperly rotated and stretched (in the preview).

This change sets all the readout_orientation to 0 by default in T19x
sensor DT files, as built-in displays are not included in development
SKUs.

Bug 2139714
Bug 1741388

Change-Id: I9ef3018e9883de2815201461fef1e56068d30cc5
Signed-off-by: Vincent Chung <vincentc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1727554
Reviewed-by: Gary Zhang <garyz@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Frank Chen <frankc@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>