t19x: dts: Update support for audio super-module
The audio super-module is supported for both Tegra194 Galen and Rey
platforms. Rey has a subset of the audio interfaces supported by Galen
and no on-board RT5658 codec. Therefore, the 'nvidia,audio-routing'
property is different between the boards and also different when the
super-module is used. Remove the 'nvidia,audio-routing' property from
the common super-module device-tree include file so that the proper
routing can be specified in the super-module device-tree include file
for each board.

Bug 2101552
Bug 200390674
Bug 200417361

Change-Id: Ifd01d9c097d9130a0e3a37dbcf6e38a833cb33af
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1735942
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>