vblk: extend the number of partitions in vblkdev
The number of partitions which can be created in vblkdev is limited by
the argument of alloc_disk(VBLK_MINORS). If more than 16 partitions are
defined in GPT, only 15 partitions can be created in the vblkdev.

This patch extends the number of partition by adding GENHD_FL_EXT_DEVT
flag in struct gendisk. In https://lwn.net/Articles/290141/, the flag is
explained like below:

There is a new flag (GENHD_FL_EXT_DEVT) which allows the disk to use
extended partition numbers; once the number of minor numbers given to
alloc_disk() is exhausted, any additional partitions will be numbered in
the extended space.

Bug 2115886

Change-Id: I687afdca61021415229c3a45606527fba0b39281
Signed-off-by: Kwangwoo Lee <kwangwool@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1738858
Reviewed-by: Phoenix Jung <pjung@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Dilan Lee <dilee@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vishal Annapurve <vannapurve@nvidia.com>
Reviewed-by: Sandeep Trasi <strasi@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>