gpu: nvgpu: Protect against possible userspace build failures
When the dev-other change to make:

  $ tmmp

Actually do something in the nvgpu repo hits dev-kernel there will
be risk of build failure. As such prevent the makefile.tmk present
in nvgpu from doing anything.

Then this build can be verified on dev-kernel with GVS (by reverting
this change) and we can ensure that no build failures are
introduced to dev-kernel.

Change-Id: I8c6fe056e12753b099927af9f0f417754edf1c1b
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1740850
Tested-by: Terje Bergstrom <tbergstrom@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>