tegra: dc: get bl framebuffer from reserved-memory
- If the display is initialized in the bootloader, the framebuffer
and lookup table will be written to a node in
/reserved-memory/fbX-carveout
- A phandle (fb_reserved) to this node is found in each nvdisplay node,
and each display will call of_tegra_get_fb_resource() to check whether
it was initialized in the bootloader, and if so it will correctly
populate the fb_mem resource field of struct tegra_dc. The node will be
accessed through the fb_reserved phandle.
- The function tegra_is_bl_display_initialized() was moved to this
file from soc/common.c, as it now no longer depends on the kernel
commandline
- If the display was not initialized, fb_mem->start will be 0
- For T210, the command line interface in common.c is preserved

Bug 2116047

Change-Id: I3391b40d4b1def8e0aa471de2e94a0d4881c7183
Signed-off-by: Axel Feldmann <afeldmann@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1741987
Reviewed-by: Mitch Luban <mluban@nvidia.com>
Tested-by: Mitch Luban <mluban@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>