hdcp: Add hdcp version information
This is a WAR to fix the HDCP regression that has creeped in due to
updation of the SRM files. As a WAR we are passing the HDCP version
information to the HDCP TA. This information will be used by the TA
to compute the correct cmac against the correct SRM file version.
The proper fix is tracked in bug
2059149.

Verified by local testing on hdcp 1.x/2.2 over hdmi and hdcp 1.x/dp
for both receiver and repeater.

Bug 200422189

Change-Id: I3a49b494d90aaea579585c63bd3ab03e09f7bd37
Signed-off-by: Pranami Bhattacharya <pranamib@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1749065
Reviewed-by: Mihir Joshi <mihirj@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Hyung Taek Ryoo <hryoo@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>