drivers: i2c: imx274: add debugfs interface
Add a debugfs 'streaming' interface to start/stop a sensor stream.
This is a port of the ov5693 interface to the imx274, they will be
refactored into a common control API once the respective drivers are
converted to the new framework style.

Also add missing mutex_destroy() to ov5693 driver.

Gerrit 1654597

Bug 200420534

Change-Id: Ia69e2cc9f9649dd2f3fdda503ae4777c560f23a6
Signed-off-by: Vincent Chung <vincentc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1751222
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>