nvlink: core: convert error print into debug print
If an nvlink device's registration succeeds, the core driver prints a
message to log this event. This is a debug print which is useful when
debugging the NVLINK driver stack. By mistake, this print was previously
being printed using nvlink_err() which meant that it was showing up
during kernel boot. This was causing a kernel_warning_test failure in
ap_systemsw. Fix this issue by using nvlink_dbg() instead of
nvlink_err() to print this debug message.

Bug 2133882
Bug 2086684
Jira NVLINK-107

Change-Id: I727a96d5ac55dd2b2557add8bbb8fb0a368b8f36
Signed-off-by: Adeel Raza <araza@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1753857
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>