video: tegra: host: Set idle EMC floors to zero
The memory bandwidth request for VIC when idle was set to max, which
is quite unnecessary, and 204MHz for other clients. It doesn't
make sense to request a non-zero EMC clock while the engine is not in
use. As such, change these to zero. It's mystery why they were
something other than zero.

Bug 200412455
Jira HOSTX-351

Change-Id: I4b82fd539a2de2f1e0ef89ca955b7c0a1daab91a
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1755712
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>