gpio: t186: Use consistent debugfs name for AON GPIOs
The debugfs node for Tegra186 GPIO AON is called 'gpio-tegra-on'
where as else in the driver the name 'gpio-tegra-aon' is used.
Furthermore, on Tegra194 for the debugfs node is called
'gpio-tegra-aon' and not 'gpio-tegra-on' which means that there
is a different name used between Tegra186 and Tegra194. For
consistency update the debugfs node name for Tegra186 to be
'tegra-gpio-aon'.

Please note that for L4T, the 'nvidia-bug-report-tegra.sh' script
reads various debugfs nodes and so having a consistent name is
necessary.

Bug 2205787

Change-Id: I53a5deafdd72b1a8d43f50edd3ac39b4b5c101b0
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1756880
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Stephen Warren <swarren@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>