ASoC: rt5659: Fix jack detection on suspend
The RT565x codec suspend/resume handlers are not called if any of the
codec DAPM widgets are powered on when entering suspend. This can
occur if a headset jack is connected which will keep on certain DAPM
widgets.

Currently, if the RT565x codec uses the JD3 pin for jack-detection
and a headset is connected, then on entering suspend the codec suspend
handler is not called on entering suspend. If the headset is
disconnected during suspend this is not detected on exiting suspend.
To workaround this is could be possible to extend the
'trigger_jack_status_check' function to trigger the jack-detection
for codec using the JD3 pin, however, a better solution is to call
the suspend/resume code for the codec from the codec's I2C
suspend/resume handlers, which will always be called. This way we
will be able remove the 'trigger_jack_status_check' and simplify
the jack detection code.

This has been tested on Tegra186 Jetson TX2 and Tegra194 Galen.

Bug 2208756

Change-Id: I8a5e70f955607af17fbc7351788dcfc70edf68ff
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1758194
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: Sameer Pujar <spujar@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Tested-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>