nvlink: Enable logging, error reporting for SYNC2x
Enable logging and error reporting for SYNC2X detected errors (Parity
and ECC) on the NVLW-MSS interface. ECC errors are separated into single
bit errors which are correctable with the ability to report when a
programmed threshold is reached and multi-bit errors which are always
fatal. Parity errors are always considered fatal.

Each error type has an associated log enable, a reporting enable and
a containment enable. This change sets each of these for fatal errors
and sets the log and report enables (but not the contain enable) for
correctable errors.

Bug 200389331

Change-Id: Ia4aa52dbf1b6c3789da2890523ca25c1f7098b69
Signed-off-by: Petlozu Pravareshwar <petlozup@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1761553
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Abhishek Sahu <absahu@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vijayakumar Subbu <vsubbu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>