thermal: throttle: fix for floorswept platforms
Pass the clock handles for all the CPU clusters instead of just the
one cluster from DT. The NAFLLL CLUSTER clock id refers to the
physical cluster and is not guaranteed to be present on floorswept
configurations.

Add logic to look for the appropriate clock handle from a list of
clock handles and fail the driver registration if a missing clock
is referenced in the cooling device.

Bug 2224577

Change-Id: I36a7d308eaa573c6b99390e3fbf36a3828377494
Signed-off-by: Srikar Srimath Tirumala <srikars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1764644
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bo Yan <byan@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>