tegra: dc: editing stub functions in header files
Certain stub functions in dc.h lead to unused definition
errors when compiled without requisite flags. Since this
is not expected behavior, changing them to inline
functions.

common.c has a hardcoded link to board-panel.h which has
functions which only get compiled when TEGRA_DC is turned
on. Since common.c does not need TEGRA_DC to be turned on,
this leads to errors on compilation if TEGRA_DC is turned
off. Hence adding stub functions in board-panel.c to
address this specific issue.

tegra_is_bl_display_initialized is onyl defined as a exported
function in common.c but not referenced in any header file.
Adding it as a stub function for now while an alternate code
path is designed which bypasses this completely.

Bug 200416228

Change-Id: I3af03cb3b3515a375cc55fd5000d3b0080d04cf3
Signed-off-by: Anshuman Kar <anshumank@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1767801
Reviewed-by: Santosh Galma <galmar@nvidia.com>
Tested-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: Naveen Kumar S <nkumars@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>