video: tegra: host: Disable SLCG WAR on K4.14
On kernel 4.14, the SLCG war is implemented by the core clock/
powergating drivers. As such, we don't need to do anything about it.
On K4.14, we also start using the PMC genpd provider for powergates,
so the assumption of nvhost_pd being the powergating provider doesn't
apply anyway.

Bug 200418280

Change-Id: I52f0e72592741d54e581780c405992a4c2e0b9e7
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1767875
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: Joseph Lo <josephl@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>