video: tegra: dc: fix seamless display on 2nd head
When bootloader intializes only the second/third head (DP in
case of Galen), seamless display would fail in kernel as nvdisplay
clock refcounting was not handled correctly. Current seamless
display logic keeps alive only the nvdisplay clock associated with
the head on which seamless is enabled. But, irrespective of the
active head, nvdisplay_p0 is needed to be enabled for common
display modules to work.

One of the solutions was to keep all the nvdisplay clocks on if
seamless was enabled on any of the heads. But, the current clock
handling in display driver doesn't disable the clocks correctly
which causes enable count imbalance.

Hence, as a WAR, replaced tegra_dc_get/put() calls with
tegra_dc_io_start/end() to eliminate clock enable and disable calls
that are not necessary. In this case, the clock enable/disable calls
around fb registration were not necessary. These calls specifically
caused nvdisplay_p0 to turn off if seamless display was not enabled
on head0 as the refcount would hit 0 during disable call.

bug 200408174

Change-Id: Ie6fe92994f5cb95ec40541238b92661107ef5bbf
Signed-off-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1771513
GVS: Gerrit_Virtual_Submit
Reviewed-by: Shu Zhong <shuz@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>